requirements.txt
pytest==8.3.4
allure-pytest==2.13.5
playwright==1.49.1
flake8-pytest-style==2.0.0